+++ tsconfig.json
@@ -15,7 +15,7 @@
 		"incremental": true,
 		"esModuleInterop": true,
 		"skipLibCheck": true,
-		"typeRoots": ["./node_modules/@types", "./src/types"]
+		"typeRoots": ["./node_modules/@types", "src/@types"]
 	},
 	"include": ["src/**/*.ts"]
 }
